Make our parser work with two types of logs.
In the previous commit, we changed our parser so that it worked only
with logs that have a source field. This mean our parser would not work
for logs with do NOT have the source field.

In this commit, the **big idea** is that we made the source field
optional. We provided a default value if no source field is found.