Parse logs in french format.
This is an exercise in doing *almost* exactly what we did in order to
parse logs that are in the US format.

The one cool thing that we did is that we derived an instace of `Enum`
for our `Product` data type, which makes it easy to write a function
with the type `Interger -> Product`.