Add error handling to our calculator.
Error handling is now necessary because our calculator performs an
operation (division) that can fail.

The key to this problem is to realize that we had to convert our pure
program `evaluate` into a monadic program, and the key to *knowing that*
is in the paper "The Essence of Functional Programming" by Philip
Wadler. To paraphrase Section 2.1 of that paper, the basic idea of
converting a pure program to monadic form is that "a function of type `a
-> b` is converted to one of type `a -> M b`".

This is exactly what we have done with our `evaluate` function. Its type
signature has changed from `Exp -> Int` to `Exp -> Eval Int`. `Eval` is
now our monadic context. Because we wanted to add error handling, and
error handling behavior is provided by the `ExceptT` monad transformer,
we defined `type Eval = Except EvalError Identity`.

Furthermore, we defined a function `runEval`. Why? Because, given the
fact that `Eval` is a monad, a value of type `Eval a` is a computation,
and we have to run the computation. As is common, our run function takes
in the entire `Eval a` value as its argument.