Tokenize uses recursion.
Previously, we went on a digression on separating out the tokenize of a
character with the traversal of a string. While this is often a great
idea, the problem here is that when we map and filter a string, we can
change the inputs shape in a constraint way.

For example, in a map, we are mapping every element of the string to a
token. This is not the behavior that we want. The input string "hello"
should be tokenized into one token, not five.