Add ability to parse multi-character identifiers.
Notice something very curious about how this works. The `tokenize`
function has to change its behavior while it is processing the input. If
the character it is processing is alphanumeric, then we have to go into
"identifier" processing mode, which is what the `identifier` function
does.

In this mode, we do not treat digits as numbers!