Parse multi-character numbers.
We use the same strategy to implement the parsing of multi-character
numbers as we did with multi-character identifiers.

What is this strategy? Namely, the first character tells us what we are
parsing - a TokenIdentifier, TokenNumber, etc. - and depending on what
we are parsing, the tokenize function "goes into a different mode". We
do this by calling a specific function - identifier, number, etc. Each
specific function, once its job is done, then goes back into "I don't
know what I am parsing mode" by calling the tokenize function. The
latter part is an example of mutual recursion.