Add the Tree data structure.
The Tree data structure is supposed to be a codification of our grammar.

There is an entire article on Wikipedia on `formal grammer`, but the
important thing to note is that for each *production* of our grammar, we
are going to have a function. Furthermore, for each function
representing a particular production, we are going to as many branches
in our code as we have rules for that production.

In the example below, we have three productions - `Expression`, `Term`,
and `Factor`. Our grammar also has *rules*. I do not know if *rule* is
the correct term, but it serves to get the point across.

```
Expression <- Term [+-] Expression
            | Identifier '=' Expression
            | Term
Term       <- Factor [*/] Term
            | Factor
Factor     <- Number
            | Identifier
            | [+-] Factor
            | '(' Expression ')'
```

Given our algorithm for codifying our grammar into code, we can expect
to have three functions - `expression`, `term`, and `factor`. For
`expression`, we can expect to have three branches. This is illustrated
in code as:

```
expression :: [Token] -> (Tree, [Token])
expression tokens =
  let (termTree, toks) = term toks
   in case lookAhead toks of
        (TokenOp op)
          | elem op [Plus, Minus] -> undefined
        TokenAssign -> undefined
        _ -> (termTree, toks)
```