Thread a Symbol Table through our evaluate function.
When we implement `evaluate` in regards to the `AssignNode`, there needs
to be a side effect of modifying a symbol lookup table, and have that
modification visible to the rest of the recursive evaluation -
particularly, when a `VarNode` looks up the symbol.

Because Haskell does not permit side-effects, we are going to thread our
symbol lookup table throughout the `evaluate` function.