Implement a parser combinator called `many`.
We can consider `many` to be a "primitive parser for repetition".
Parsers that utilize recursion / repetition, such as `word` and
`digits`, can be implemented in terms of `many`.

The concept of repetition / recursion in parsing is very important -
oftentimes, we can't predetermine how many items of a particular type we
have.