Prepare to use Validation data type.
Although the Validation datatype is isomorphic to Either, it has a
different instance of Applicative that allows it to accumulate errors.
Either, if you'll remember, returns the first error that it encounters.

Like Either, Validation is parameterized over two type variables. In
order to have an instance of Applicative, the first type variable must
have an instance of Semigroup, so that that errors can accumulate.
That's why we are changing the definition of Error.