Refactor our code to use `Validation`.
Remember that `Validation` is isomorphic to `Either`. The reason we have
`Validation` when we already have `Either` is that the `Applicative`
instance for `Validation e` **accumulates** `Failure` values.

In order to make use of `Validation e`'s accumulating behavior, we have
to use the *sparrow* operators from the `Applicative` typeclass. I have
never used these operators before. What I learned is that the sparrow
operators ignore one of the values to which they are applied to, but,
*crucially*, the sparrow operators do not ignore the effects of their
operands. Like `Either`, we often use `Validation` to keep track of
failure. The sparrow operators, for `Validation e`, will ignore one of
the values **if** the value does not produce any errors, but it will
concatenate the results of the "ignored" value with the "non-ignore"
value if there are errors. This, for me at least, is the key takeaway of
this book.